Move HamburgerButton outside ModuleHeader; pass it as `children`
