+++ app/src/editor/shared/components/HamburgerButton/hamburgerButton.pcss
@@ -4,6 +4,7 @@
   background: none;
   border: 0;
   box-sizing: content-box;
+  flex-shrink: 0;
   height: 24px;
   padding: 8px;
   width: 24px;
