+++ app/src/editor/shared/components/ModuleHeader/moduleHeader.pcss
@@ -24,10 +24,6 @@
   width: 24px;
 }
 
-.menuToggle {
-  flex-shrink: 0;
-}
-
 .dragHandle,
 .dragCancel {
   /* Class names for react-draggable. */
