+++ app/src/editor/shared/components/ModuleHeader/moduleHeader.stories.jsx
@@ -19,7 +19,6 @@ stories.add('default', () => (
             <ModuleHeader
                 label={label}
                 onLabelChange={setLabel}
-                onHamburgerClick={() => {}}
             />
         )}
     </UseState>
